# swagger_file.yaml
openapi: 3.0.0
info:
  title: Todo API
  description: API for managing todos
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /todos:
    get:
      summary: Get all todos
      responses:
        '200':
          description: A list of todo items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    completed:
                      type: boolean
    post:
      summary: Create a new todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                completed:
                  type: boolean
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  completed:
                    type: boolean
  /todos/{id}:
    put:
      summary: Update an existing todo
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the todo to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                completed:
                  type: boolean
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  completed:
                    type: boolean
    delete:
      summary: Delete a todo
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the todo to delete
      responses:
        '204':
          description: Todo deleted